xfree86: Remove ScrnInfoRec::imageByteOrder
No driver ever overrode this.  It can't ever have worked very well to
disagree with the rest of the server, as it's a property of the wire
encoding, meaning it can't vary among screens.  We may as well just nail
it down.

I do like how the error message here doesn't match what it's checking,
though.  Nice touch.

Signed-off-by: Adam Jackson <ajax@redhat.com>