xfree86: Remove 24bpp pixmap format support
There's really no reason to pretend to support this, apps hate it, all
we're doing is giving people a way to injure themselves.

I think this might sorta break cirrus?  In that it looks like it'll
prefer 24bpp (which I think was my doing in the first place), and claims
not to be able to support 32bpp pixmaps when that's the case.  I think
that's really a comment about how XAA worked, fb certainly has no
trouble with the notion.  At any rate the affected chips should probably
default to 16 instead, since cirrus cards are usually low on VRAM to
begin with.

Signed-off-by: Adam Jackson <ajax@redhat.com>