xfree86: Remove leaky /dev/mem checks from xf86OSInitVidMem
This is mostly a no-op, the checks didn't have much effect since
pciaccess didn't end up using the fd we opened.  Except for OpenBSD,
where you have to pass that in from above, which is sort of a weird API
decision.  So this will break there until libpciaccess is made less
dumb.

Signed-off-by: Adam Jackson <ajax@redhat.com>