xfree86: Remove various "weak" functions
I added these way back in the dlloader conversion.  The bug at the time
was that some drivers had data relocations for these functions (ie,
mentioned them as function pointers not as calls), and would thus fail
to load because the providing module (fbdevhw, vgahw, or shadow) wasn't
loaded yet and you can't lazily resolve data relocs.

It works but it's ugly.  The driver can just as easily include its own
functions that do nothing but call the server code, which is also
portable to servers that do provide the weak functions.  Yes this is
deleting code from xserver at the expense of adding more code to 18
drivers, but drivers are already ugly.

Signed-off-by: Adam Jackson <ajax@redhat.com>