hw: Remove LD_EXPORT_SYMBOLS_FLAG from servers that don't need it
Which is most of them.  This used to be necessary to get the glapi
dispatch exported from the server binary, but that's not the case
anymore.  (Weird then that Xnest had this, since it has no GLX support;
and that Xdmx had this, since it has its own GLX implementation).
xfree86 still needs it for driver ABI though.

Signed-off-by: Adam Jackson <ajax@redhat.com>