render: Remove the XID from the PictureRec
This isn't something the driver needs to know.

Probably this is a little broken?  I think the only thing the
FreeResource/FreePicture split could do is, in the Xinerama case,
destroy the XRT_PICTURE before the Picture.  But that should be
harmless, right?  XineramaDeleteResource is just free().  Anyway, test
this theory.

Signed-off-by: Adam Jackson <ajax@redhat.com>