pseudoramix: Add an id parameter to tracked screens
This parameter is opaque from pseudoramix' perspective, it merely exists
to uniquely identify a display.  This is in preparation for hotplug
scenarios when we'll want to add and remove individual screens.

For the moment in xquartz this is the CGDirectDisplayID (just an int32).
But when randr is ported to pseudoramix this will likely become the
pointer to the crtc.

Signed-off-by: Adam Jackson <ajax@redhat.com>