glamor: Accelerate XYBitmap PutImage
XYBitmap to a drawable of depth != 1 is special. Where bits are 1, we
draw the foreground color; where they are 0, we draw the background
color. To handle this without falling back, allocate a temporary pixmap,
do the color expansion in software, and then upload that image to the
destination.

There's a few other ways to skin this cat (add an R1 texture format to
GL, convert that image to a stipple and loop back through FillRect...)
but this is quite fast enough.

Signed-off-by: Adam Jackson <ajax@redhat.com>