dix: Move {Change,Copy,Destroy}Clip from GCFuncs to Screen
There are only two behaviors here. Xnest and Xdmx forward the clip list
to the backing server, other DDXes use mi. But a layer like damage has
to wrap everything in the GCFuncs vtable even though it really only
cares about ValidateGC.

Since none of this varies within a screen, we can move the clip
management funcs up to the ScreenRec and eliminate the empty wrappers.

Signed-off-by: Adam Jackson <ajax@redhat.com>