Fix linking xfree86 power management support
This is insanely ugly but it works. The problem is, none of the symbols
defined by xf86PM.c are referenced before libcommon.a is searched, but
likewide none of the symbols defined by the OS PM code are referenced by
the time libxorgos is searched. Just fold xf86PM into xf86Helper (which
does satisfy references by that point of the link) and move on.

The interface between the xf86 and os layers here is quite ugly and this
could probably be done more cleanly, should someone feel motivated.

Signed-off-by: Adam Jackson <ajax@redhat.com>